+++ src/cinema6-jsonp.js
@@ -261,7 +261,9 @@
             title: experience.data.title,
             container: ctr,
             context: context,
-            adId: adId
+            adId: adId,
+            startPixel: params.startPixel,
+            countPixel: params.countPixel
         };
     }