+++ test/spec/cinema6-jsonp.ut.js
@@ -443,6 +443,44 @@ describe('cinema6-jsonp.js', function() {
                 });
             });
 
+            describe('if a startPixel is specified', function() {
+                beforeEach(function(done) {
+                    c6.embeds.length = 0;
+                    load(function() {
+                        adtech.config.placements['108542'].complete();
+                        waitForDeps(expIds.map(function(id) {
+                            return baseUrl + '/api/public/content/experience/' + id + '.js?container=jsonp';
+                        }), done);
+                    }, '/base/src/cinema6-jsonp.js?callback=onC6AdLoad&id=108542&startPixel=http%3A%2F%2Ftracking.com%2Fpixel&cb=' + Date.now());
+                });
+
+                it('should set the startPixel on the config', function() {
+                    expect(c6.embeds.length).toBe(expIds.length);
+                    c6.embeds.forEach(function(embed) {
+                        expect(embed.config.startPixel).toBe('http://tracking.com/pixel');
+                    });
+                });
+            });
+
+            describe('if a countPixel is specified', function() {
+                beforeEach(function(done) {
+                    c6.embeds.length = 0;
+                    load(function() {
+                        adtech.config.placements['108542'].complete();
+                        waitForDeps(expIds.map(function(id) {
+                            return baseUrl + '/api/public/content/experience/' + id + '.js?container=jsonp';
+                        }), done);
+                    }, '/base/src/cinema6-jsonp.js?callback=onC6AdLoad&id=108542&countPixel=http%3A%2F%2Ftracking.com%2Fpixel&cb=' + Date.now());
+                });
+
+                it('should set the startPixel on the config', function() {
+                    expect(c6.embeds.length).toBe(expIds.length);
+                    c6.embeds.forEach(function(embed) {
+                        expect(embed.config.countPixel).toBe('http://tracking.com/pixel');
+                    });
+                });
+            });
+
             describe('after the experiences have been fetched', function() {
                 var exps;
 
@@ -546,7 +584,9 @@ describe('cinema6-jsonp.js', function() {
                             title: exp.data.title,
                             container: 'jsonp',
                             context: 'jsonp',
-                            adId: undefined
+                            adId: undefined,
+                            startPixel: undefined,
+                            countPixel: undefined
                         });
                     });
                 });